Added Fling Settings
Added capability of letting the pan gesture "fling" the menu open or closed based on variables the user can set, namely the FlingPercentage and FlingVelocity. The FlingPercentage determines the percentage of the menu's width that a slide gesture can open or close the menu which will autocomplete the action, even if velocity falls to 0. The FlingVelocity determines how fast a slide can be to fling the menu open regardless what X position it is at when the gesture ends.

These settings were added to make the menu fall in line with other iOS app common practices, and namely help with sidebar integration for the ParkingPanda iOS app.